CRUD Category

+++ NewManagementSystem/NewsManagementSystem.BLL/Services/Category/ICategoryService.cs
@@ -6,11 +6,11 @@ public interface ICategoryService
 {
     Task<List<Category>> GetCategoriesAsync();
     Task<Category?> GetCategoryByNameAsync(string name);
-    Task<Category?> GetCategoryByIdAsync(short categoryId);
+    Task<Category?> GetCategoryByIdAsync(int categoryId);
     Task CreateCategoryAsync(Category category);
     Task UpdateCategoryAsync(Category category);
     Task DeleteCategoryAsync(Category category);
-    Task<bool> CategoryExistsAsync(short categoryId);
+    Task<bool> CategoryExistsAsync(int categoryId);
     Task<List<Category>> SearchCategoriesByNameAsync(string searchTerm);
 
 }
